refactor: slim RNG checkpoint review surface
Move runner checkpoint helpers into checkpointing.py and hash SQLite tables instead of per-column watermarks so the PR stays smaller without changing resume behavior.
